requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]

# Database
sqlalchemy
alembic
psycopg2-binary

# Authentication and Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Configuration and Validation
pydantic[email]
pydantic-settings
python-dotenv

# Rate Limiting and Caching
redis
slowapi
msgpack

# Testing
pytest
pytest-asyncio
httpx

# Development Tools
black
isort
flake8
click